refactor: clean up TUI and agent tools
Amp-Thread-ID: https://ampcode.com/threads/T-019e5d28-f2e1-76cc-a529-287e4646a616
